requirements.txt
Flask
Flask-SocketIO
flask-cors
eventlet
opencv-python
mediapipe
numpy
Pillow
PyMuPDF
pyautogui
pynput
python-dotenv
